+++ src/main/java/org/idnt/udemy/springboot/app/example/repository/BankRepository.java
@@ -1,0 +1,11 @@
+package org.idnt.udemy.springboot.app.example.repository;
+
+import org.idnt.udemy.springboot.app.example.model.Bank;
+
+import java.util.List;
+
+public interface BankRepository {
+    List<Bank> findAll();
+    Bank findById(Long id);
+    void update(Bank account);
+}